Display team tricks in Score component. Change names of Score and ScoreBoards components
