+++ app/js/components/ScoreBoard.js
@@ -1,47 +1,76 @@
-import React, {useContext} from 'react';
-import {connect} from 'react-redux';
-import {selectTricks, selectPlayers} from '../redux/selectors';
-import {newGame} from '../redux/actions';
+import React, { useEffect, useContext } from 'react';
+import { connect } from 'react-redux';
+import { createStructuredSelector } from 'reselect';
+import { getScore } from '../redux/actions';
 import { LocalStateContext } from '../pages/GamePage.js';
+import {selectScore,
+				selectPlayers,
+				selectLastMasterIndex,
+				selectCurrentPlayer,
+				selectCurrentDeclaration,
+				selectTeams,
+				selectPartnerId } from '../redux/selectors';
 
-import Score from './Score';
-import Card from './Card';
+const ScoreBoard = ({ getScore, score, players, currentPlayer, lastMasterIndex, currentDeclaration, teams, partnerId }) => {
 
-const ScoreBoard = ({players, tricks, newGame}) => {
+	const {tableId} = useContext(LocalStateContext);
 
-  const {tableId} = useContext(LocalStateContext);
-  return (
-    <div className="has-text-centered">
-      <Score />
-      {
-        players.map(p => {
-          const playerTricks = tricks.filter(({playerIndex}) => p.index === playerIndex)
-          return (
-            <div key={p.index}>
-              <h2>{p.name || 'BOT'}</h2>
-              <div className="hand">
-              {
-                playerTricks.map(({cards}) => {
-                  return (cards.length) ? cards.map(c => <Card key={c} value={c} />) : <div>Vous n'avez fait aucun plis</div>
-                })
-              }
-              </div>
-            </div>
-          );
-        })
-      }
-      <button className="button is-primary is-large" onClick={() => newGame(tableId)} style={{marginTop: '2rem'}}>New game</button>
-    </div>
-  );
+	useEffect(() => {
+	  getScore(tableId)
+	}, []);
+
+	return (
+		<div className="table-container">
+			<table className="table is-fullwidth">
+			<tr>
+				<td>
+					<table className="table is-fullwidth">
+						<thead><tr><td>&nbsp;</td></tr></thead>
+						<tbody>
+							<tr>Points</tr>
+							<tr>Dix de der</tr>
+							<tr>Belote</tr>
+							<tr>Total</tr>
+							<tr>Total de la partie</tr>
+						</tbody>
+					</table>
+				</td>
+				{
+					teams.map( (team) => {
+						return (!team.currentGame) ?
+							null :
+							<td key={team.name}>
+								<table className="table is-fullwidth">
+									<thead><tr><td>{team.name}</td></tr></thead>
+									<tbody>
+										<tr>{team.currentGame.gameScore}</tr>
+										<tr>{(team.currentGame.hasLastTen) ? 10 : 0}</tr>
+										<tr>{(team.currentGame.hasBelote) ? 20 : 0}</tr>
+										<tr>{team.currentGame.gameTotal}</tr>
+										<tr>{(team.totalScore || 0) + team.currentGame.gameTotal}</tr>
+									</tbody>
+								</table>
+							</td>
+					})
+				}
+			</tr>
+			</table>
+		</div>
+	);
 };
 
 const mapDispatchToProps = (dispatch) => ({
-  newGame: (tableId) => dispatch(newGame(tableId)),
+	getScore: (tableId) => dispatch(getScore(tableId)),
 });
 
-const mapStateToProps = (state) => ({
-  players: state.players,
-  tricks: state.tricks,
-})
+const mapStateToProps = createStructuredSelector({
+	score: selectScore,
+	players: selectPlayers,
+	lastMasterIndex: selectLastMasterIndex,
+	currentPlayer: selectCurrentPlayer,
+	currentDeclaration: selectCurrentDeclaration,
+	teams: selectTeams,
+	partnerId: selectPartnerId
+});
 
 export default connect(mapStateToProps, mapDispatchToProps)(ScoreBoard);
